newsletter_bot/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.7
openai==1.3.7
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
apscheduler==3.10.4
feedparser==6.0.10